add clang tidy analyzer support, and restructure result handling, new modules store the results to database or print them to the output
