+++ codechecker_lib/analyzers/result_handler_clang_tidy.py
@@ -9,35 +9,48 @@ import ntpath
 import shared
 
 from codechecker_lib import logger
+from codechecker_lib import tidy_output_converter
+
+from codechecker_lib.analyzers import result_handler_store_db
+from codechecker_lib.analyzers import result_handler_printout
+
 from codechecker_lib.analyzers import result_handler_base
 
 LOG = logger.get_new_logger('CLANG_TIDY_RESULT_HANDLER')
 
 
-class CTDBResHandler(result_handler_base.ResultHandler):
+class CTDBResHandler(result_handler_store_db.ResultHandlerStoreToDB):
     """
     Clang tidy results database handler
     """
 
-    def __init__(self, buildaction, workspace, run_id):
-        super(CTDBResHandler, self).__init__(buildaction, workspace)
-        self.__connection = connection
-        self.__workspace = workspace
-        self.__run_id = run_id
-
     def postprocess_result(self):
-        pass
+        """
 
-    def handle_results(self):
-        pass
+        """
+        parser = tidy_output_converter.OutputParser()
+
+        messages = parser.parse_messages(self.analyzer_stdout.splitlines())
+
+        plist_converter = tidy_output_converter.PListConverter()
+        plist_converter.add_messages(messages)
+
+        plist = self.get_analyzer_result_file()
+        plist_converter.write_to_file(plist)
 
 
-class CTQCResHandler(result_handler_base.ResultHandler):
+class CTQCResHandler(result_handler_printout.ResultHandlerPrintOut):
     """
+
     """
 
     def postprocess_result(self, result):
-        pass
+        parser = tidy_output_converter.OutputParser()
 
-    def handle_results(self):
-        pass
+        messages = parser.parse_messages(self.analyzer_stdout.splitlines())
+
+        plist_converter = tidy_output_converter.PListConverter()
+        plist_converter.add_messages(messages)
+
+        plist = self.get_analyzer_result_file()
+        plist_converter.write_to_file(plist)
